backend/requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.30
SQLAlchemy~=2.0
psycopg[binary]~=3.1
alembic~=1.13
PyJWT~=2.9
python-dotenv~=1.0
pydantic~=2.8
pydantic-settings~=2.4
kubernetes~=29.0
tenacity~=8.2
passlib~=1.7
httpx~=0.27
email-validator~=2.2
PyYAML~=6.0
